refactor: 💡 Refactor file naming and locale choosing logic
