+++ apps/webstore/src/app/pages/products/mini-product-display/mini-product-display.component.ts
@@ -1,5 +1,5 @@
 import { Component, Input } from '@angular/core';
-import { ProductsItem } from '../productsItems';
+import { ProductsItem } from '../products-items.constant';
 
 @Component({
   selector: 'webstore-mini-product-display',
