+++ apps/webstore/src/app/pages/products/product-display/product-display.component.spec.ts
@@ -1,5 +1,5 @@
 import { ComponentFixture, TestBed } from '@angular/core/testing';
-import { ProductsItem } from '../productsItems';
+import { ProductsItem } from '../products-items.constant';
 
 import { ProductDisplayComponent } from './product-display.component';
