+++ apps/webstore/src/app/pages/products/products.component.ts
@@ -1,9 +1,7 @@
 import { Component } from '@angular/core';
-import { ProductsItem } from './productsItems';
-import productsItems from './productsItems';
 import { LocalStorageService } from '@local-storage';
 import { LocaleType, LocalStorageVars } from '@models';
-import { BehaviorSubject } from 'rxjs';
+import productsItems, { ProductsItem } from './products-items.constant';
 
 @Component({
   selector: 'webstore-products',
@@ -13,15 +11,13 @@ import { BehaviorSubject } from 'rxjs';
 export class ProductsComponent {
   productsList: ProductsItem[] = productsItems;
 
-  locale$: BehaviorSubject<LocaleType>;
   localeCode: LocaleType;
 
   constructor(private localStorage: LocalStorageService) {
-    this.locale$ = this.localStorage.getItem<LocaleType>(LocalStorageVars.locale);
-    this.localeCode = this.locale$.getValue();
+    this.localeCode = this.localStorage.getItem<LocaleType>(LocalStorageVars.locale).value;
   }
 
   isEnglish(): boolean {
-    return this.localeCode === 'en-US';
+    return this.localeCode === LocaleType.en;
   }
 }
